Updates based on 2nd round of review (part 4).
Adding clarification to proof of Theorem 4.1; changing
sentence that started "In a perturbation, we replace ..." so that
it's clear what is being referred to. I intended "a perturbation" to
refer to the difference `dF` in `F + dF` but the reviewer read it
as all of `F + dF`. So I changed it to "In a perturbed `F` ..." and
replaced `delta_j` with `c_j + delta_j`.